+++ NEWS.md
@@ -1,4 +1,9 @@
 # 0.1.2
+## New features
+* NMreadExt is a new function that reads parameter estimates,
+  uncertainties if available, estimation iterations and other
+  information from .ext files. 
+
 ## Other improvements
 * NMreadCsv supports multiple prioritized formats in the `format`
   arguments.